resource IDs as uint32_t instead of uint64_t

+++ resources/ntshengn_resources_audio.h
@@ -6,7 +6,7 @@
 namespace NtshEngn {
 
 	// Sound
-	typedef uint64_t SoundID;
+	typedef uint32_t SoundID;
 
 	struct Sound {
 		// Number of channels (1 = Mono, 2 = Stereo, ...)
